requirements.txt
numpy>=1.19.2
tensorflow>=2.4.0
scikit-learn>=0.24.0
Pillow>=8.0.0
joblib>=1.0.0
flask>=2.0.0
gunicorn>=20.1.0 